# doc/api.yml
openapi: 3.0.3
info:
  title: jwt4poja API
  version: 1.0.0
  description: |
    Grade management API for a 3-year L2 program (POJA-generated Spring Boot
    app, package `com.techindna.anerti`, schema `jwt4poja_app`).

    Roles: `ADMIN`, `TEACHER`, `STUDENT`. There is no public registration:
    accounts are created by an `ADMIN` (`POST /students`, `POST /teachers`).
    Auth endpoints (`/auth/**`) are public; everything else requires a JWT.
servers:
  - url: http://localhost:8080
...
paths:
  /auth/login:
    post:
      tags: [auth]
      summary: Login and request a verification link
      description: Checks credentials and sends a login-verification email.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        '202':
          description: Accepted — login-verification email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageBody'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableContent'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/verification/{token}:
    get:
      tags: [auth]
      summary: Verify the email token and obtain a JWT
      description: |
        Single-use token (15-minute TTL). Consumes the token and returns a JWT
        plus the user payload. Serves login verification.
      operationId: verify
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Verified — JWT issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyRegistrationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /teachers:
    post:
      tags: [teachers]
      summary: Create a teacher (ADMIN)
      description: |
        Creates the underlying user account (role `TEACHER`) and its teacher
        inheritance record in one operation. `username`, `email` and `ref` must
        be unique.
      operationId: createTeacher
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeacherInput'
      responses:
        '201':
          description: Teacher created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserExtendTeacher'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableContent'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags: [teachers]
      summary: List teachers (ADMIN)
      description: |
        Lists teachers. `search` matches `ref`, first name or last name
        (case-insensitive substring, partial match). `teacherStatus` filters by
        exact status.
      operationId: listTeachers
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: teacherStatus
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/TeacherStatus'
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Page of teachers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableContent'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from `GET /auth/verification/{token}` (subject = user id, claim `role`)
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          example:
            status: 400
            error: Bad Request
            message: "Malformed request: invalid UUID in path"
            timestamp: 2026-08-13T12:00:00Z
    Unauthorized:
      description: Missing/invalid JWT or bad credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          example:
            status: 401
            error: Unauthorized
            message: Missing or invalid token
            timestamp: 2026-08-13T12:00:00Z
    Forbidden:
      description: Authenticated but not allowed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          example:
            status: 403
            error: Forbidden
            message: Access denied for this role
            timestamp: 2026-08-13T12:00:00Z
...
    Conflict:
      description: Uniqueness conflict (username, email or ref already in use)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          example:
            status: 409
            error: Conflict
            message: A resource with this unique value already exists
            timestamp: 2026-08-13T12:00:00Z
    UnprocessableContent:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          example:
            status: 422
            error: Unprocessable Content
            message: "Validation failed: coefficient must be greater than 0"
            timestamp: 2026-08-13T12:00:00Z
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          example:
            status: 500
            error: Internal Server Error
            message: An unexpected error occurred
            timestamp: 2026-08-13T12:00:00Z
  schemas:
    UserRole:
      type: string
      enum: [ADMIN, TEACHER, STUDENT]
...
    TeacherStatus:
      type: string
      enum: [ACTIVE, INACTIVE, OTHER]
...
    User:
      type: object
      required: [id, username, firstName, lastName, email, role, createdAt]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
          maxLength: 50
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 100
        role:
          $ref: '#/components/schemas/UserRole'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
      example:
        id: 3f3e9b2a-1c4e-4a9d-8f2b-5e7a1c0d9f21
        username: mdupont
        firstName: Marie
        lastName: Dupont
        email: marie.dupont@hacheuil.edu
        role: STUDENT
        createdAt: 2023-09-04T09:00:00Z
        updatedAt: null
    LoginInput:
      type: object
      required: [password]
      description: Either `username` or `email` plus `password`
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
      example:
        username: mdupont
        password: s3cur3P@ss
    MessageBody:
      type: object
      required: [message]
      properties:
        message:
          type: string
      example:
        message: A verification link has been sent to your email
    VerifyRegistrationResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
          description: JWT (subject = user id, claim `role`)
        user:
          $ref: '#/components/schemas/User'
      example:
        token: eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIzZjNlOWIyYSIsInJvbGUiOiJTVFVERU5UIn0.SIGNATURE
        user:
          id: 3f3e9b2a-1c4e-4a9d-8f2b-5e7a1c0d9f21
          username: mdupont
          firstName: Marie
          lastName: Dupont
          email: marie.dupont@hacheuil.edu
          role: STUDENT
          createdAt: 2023-09-04T09:00:00Z
          updatedAt: null
...
    TeacherInheritance:
      type: object
      required: [id, ref, teacherStatus]
      properties:
        id:
          type: string
          format: uuid
          description: Teacher inheritance id
        ref:
          type: string
          maxLength: 10
        joinedAt:
          type: string
          format: date-time
          nullable: true
        teacherStatus:
          $ref: '#/components/schemas/TeacherStatus'
      example:
        id: 5c3d4e5f-6a7b-4c8d-9e0f-1a2b3c4d5e6f
        ref: T001
        joinedAt: 2023-09-01T08:00:00Z
        teacherStatus: ACTIVE
    UserExtendTeacher:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          required: [teacherInheritance]
          properties:
            teacherInheritance:
              $ref: '#/components/schemas/TeacherInheritance'
      description: User account with its embedded teacher inheritance record.
      example:
        id: 2f4e9b2a-1c4e-4a9d-8f2b-5e7a1c0d9f21
        username: p.martin
        firstName: Paul
        lastName: Martin
        email: paul.martin@hacheuil.edu
        role: TEACHER
        createdAt: 2023-09-01T08:00:00Z
        updatedAt: null
        teacherInheritance:
          id: 5c3d4e5f-6a7b-4c8d-9e0f-1a2b3c4d5e6f
          ref: T001
          joinedAt: 2023-09-01T08:00:00Z
          teacherStatus: ACTIVE
    CreateTeacherInput:
      type: object
      required: [username, password, firstName, lastName, email, ref]
      description: Creates the user account (role TEACHER) and its teacher inheritance record.
      properties:
        username:
          type: string
          maxLength: 50
        password:
          type: string
          format: password
          minLength: 12
          description: 12+ chars with upper, lower, digit and special char
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 100
        ref:
          type: string
          maxLength: 10
        joinedAt:
          type: string
          format: date-time
          nullable: true
        teacherStatus:
          $ref: '#/components/schemas/TeacherStatus'
      example:
        username: p.martin
        password: s3cur3P@ss
        firstName: Paul
        lastName: Martin
        email: paul.martin@hacheuil.edu
        ref: T001
        joinedAt: 2023-09-01T08:00:00Z
        teacherStatus: ACTIVE
...
    Meta:
      type: object
      required: [page, size, total]
      properties:
        page:
          type: integer
          minimum: 1
        size:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          format: int64
      example:
        page: 1
        size: 10
        total: 1
...
    TeacherListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/UserExtendTeacher'
        meta:
          $ref: '#/components/schemas/Meta'
...
    ErrorBody:
      type: object
      required: [status, error, message, timestamp]
      properties:
        status:
          type: integer
          description: HTTP status code
        error:
          type: string
          description: Uppercased reason phrase, e.g. CONFLICT
        message:
          type: string
        timestamp:
          type: string
          format: date-time